Move everything ledger related to ledger.ts file
Removes `src/lib/utils/ledger` folder and moves everything to
`src/lib/utils/ledger.ts` file.
