--- typescript/src/lib/utils/ledger/index.ts
@@ -1,2 +1,0 @@
-export * from "./ledger"
-export * from "./wallet-api"
