--- typescript/src/lib/utils/ledger/wallet-api.ts
@@ -1,16 +1,0 @@
-import {
-  WalletAPIClient,
-  WindowMessageTransport,
-} from "@ledgerhq/wallet-api-client"
-
-export const getWindowMessageTransport = () => {
-  return new WindowMessageTransport()
-}
-
-export const getWalletAPIClient = (
-  windowMessageTransport: WindowMessageTransport
-) => {
-  const walletApiClient = new WalletAPIClient(windowMessageTransport)
-
-  return walletApiClient
-}
